Added external embeds to the report command.

+++ commands/report.js
@@ -1,35 +1,25 @@
-const Discord = require('discord.js'); // this links to the official Discord npm package
-const config = require('../config.json'); // this links to the config.json file
+const Discord = require('discord.js');
+const config = require('../config.json');
+const errors = require('../util/errors.js');
 
 module.exports.run = async (client, message, args) => {
-  let user = message.guild.member(message.mentions.users.first() || message.guild.members.get(args[0]));
-  if (!user) {
-    const embed = new Discord.RichEmbed()
-    .setTitle('An error has occurred!')
-    .setColor(config.red)
-    .setDescription('The user you have requested to punish could not be found or a reason has not been supplied for this report.');
-    message.channel.send(embed);
-    message.delete().catch(O_o=>{});
-  };
+  let user = message.guild.member(message.mentions.users.first()) || message.guild.members.get(args[0]);
+  if (!user) return errors.invalidUser(message);
+
   let reason = args.join(' ').slice(22);
+  if (!reason) return errors.invalidReason(message);
 
   let embed = new Discord.RichEmbed()
   .setTitle('Incoming Report!')
-  .setColor(config.red)
-  .addField('Reported User', `${user} with ID: ${user.id}`)
-  .addField('Reported By:', `${message.author} with ID: ${message.author.id}`)
-  .addField('Reported in Channel:', message.channel)
-  .addField('Time:', message.createdAt)
-  .addField('Reason:', reason);
+  .setColor(config.yellow)
+  .addField('Reported User', `${user}`)
+  .addField('Reported By', `${message.author}`)
+  .addField('Reported in Channel', message.channel)
+  .addField('Time', message.createdAt)
+  .addField('Reason', reason);
 
   let reportschannel = message.guild.channels.find('name', 'reports');
-  if (!reportschannel) {
-  const embed = new Discord.RichEmbed()
-  .setTitle('An error has occurred!')
-  .setColor(config.red)
-  .setDescription('A `#reports` channel could not be found!\nPlease contact your guild/server Administrator to create one.');
-  message.channel.send(embed);
-  };
+  if (!reportschannel) return errors.noReportChannel(message);
 
   message.delete().catch(O_o=>{});
   reportschannel.send(embed);
